app/requirements.txt
asgiref==3.8.1
Django==5.1.1
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
sqlparse==0.5.1
celery==5.4.0
redis==5.1.1
flower==2.0.1
django_celery_beat==2.7.0
django-redis==5.4.0
djangorestframework==3.15.2
markdown==3.7
django-filter==24.3